pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-kit"
version = "0.2.0"
description = "4-layer RAG skeleton (Data / Retriever / Generator / Pipeline)"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "apps*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
